+++ routes/Volunteer.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { signupVolunteer, loginVolunteer, getUserDetails, updateUserProfile } = require('../controllers/products');
+const { signupVolunteer, loginVolunteer, getUserDetails, updateUserProfile } = require('../controllers/Volunteer');
 const auth = require('../middleware/auth');
 const multer = require('multer');
 const path = require('path');
